Moved LoadDetail and its adder in public API
