+++ include/powsybl/iidm/extensions/LoadDetail.hpp
@@ -5,8 +5,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-#ifndef POWSYBL_IIDM_EXTENSIONS_IIDM_LOADDETAIL_HPP
-#define POWSYBL_IIDM_EXTENSIONS_IIDM_LOADDETAIL_HPP
+#ifndef POWSYBL_IIDM_EXTENSIONS_LOADDETAIL_HPP
+#define POWSYBL_IIDM_EXTENSIONS_LOADDETAIL_HPP
 
 #include <powsybl/iidm/AbstractMultiVariantIdentifiableExtension.hpp>
 
@@ -17,8 +17,6 @@ namespace iidm {
 class Load;
 
 namespace extensions {
-
-namespace iidm {
 
 class LoadDetail : public AbstractMultiVariantIdentifiableExtension {
 public:  // Extension
@@ -98,12 +96,10 @@ private:
     std::vector<double> m_variableReactivePower;
 };
 
-}  // namespace iidm
-
 }  // namespace extensions
 
 }  // namespace iidm
 
 }  // namespace powsybl
 
-#endif  // POWSYBL_IIDM_EXTENSIONS_IIDM_LOADDETAIL_HPP
+#endif  // POWSYBL_IIDM_EXTENSIONS_LOADDETAIL_HPP
