+++ include/powsybl/iidm/extensions/LoadDetailAdder.hpp
@@ -5,8 +5,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-#ifndef POWSYBL_IIDM_EXTENSIONS_IIDM_LOADDETAILADDER_HPP
-#define POWSYBL_IIDM_EXTENSIONS_IIDM_LOADDETAILADDER_HPP
+#ifndef POWSYBL_IIDM_EXTENSIONS_LOADDETAILADDER_HPP
+#define POWSYBL_IIDM_EXTENSIONS_LOADDETAILADDER_HPP
 
 #include <powsybl/iidm/ExtensionAdder.hpp>
 
@@ -15,8 +15,6 @@ namespace powsybl {
 namespace iidm {
 
 namespace extensions {
-
-namespace iidm {
 
 class LoadDetailAdder : public ExtensionAdder {
 public:
@@ -106,12 +104,10 @@ private:
     double m_variableReactivePower = 0.0;
 };
 
-}  // namespace iidm
-
 }  // namespace extensions
 
 }  // namespace iidm
 
 }  // namespace powsybl
 
-#endif  // POWSYBL_IIDM_EXTENSIONS_IIDM_LOADDETAILADDER_HPP
+#endif  // POWSYBL_IIDM_EXTENSIONS_LOADDETAILADDER_HPP
